Added lux::core::c::shaders::updateShader function. Added clear shader update in swapchain recreation. Fixed lux::Map "isValid" function

+++ LuxEngine_VsProject/LuxEngine/Types/EngineTypes.h
@@ -7,7 +7,7 @@
 
 
 
-typedef uint32 LuxShader, LuxBuffer;
+typedef uint64 LuxShader, LuxBuffer;
 typedef uint64 LuxCell;
